tests/wpt/mozilla/tests/mozilla/child_reparenting.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body><font><p></font>
<script>
test(function() {
  assert_equals(document.body.firstChild.tagName, "FONT");
  assert_equals(document.body.firstChild.nextSibling.tagName, "P");
});
</script>
</body>
</html>
